+++ include/Camera.h
@@ -41,16 +41,15 @@ class Camera {
     static void setRoom(int r) { if (room != r) dirty = true; room = r; }
     static int getRoom() { return room; }
 
-    static void setShowOverlay(bool s) { showOverlay = s; }
-    static bool getShowOverlay() { return showOverlay; }
-
     static void setKeepInRoom(bool k) { keepInRoom = k; }
     static bool getKeepInRoom() { return keepInRoom; }
 
     static bool boxInFrustum(BoundingBox b);
     static void displayFrustum(glm::mat4 MVP);
 
-    static void displayUI();
+    static const float fov;
+    static const float nearDist;
+    static const float farDist;
 
   private:
     static void calculateFrustumPlanes();
@@ -62,7 +61,7 @@ class Camera {
     static glm::mat4 projection;
     static glm::mat4 view;
     static float rotationDeltaX, rotationDeltaY;
-    static bool updateViewFrustum, dirty, showOverlay, movingFaster;
+    static bool updateViewFrustum, dirty, movingFaster;
     static bool keepInRoom;
     static int room;
 };